feat: add request handler and user validation
